Personal_Meeting_Agent1/Meeting_Agent1/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.31.0
python-dateutil==2.8.2
pydantic>=2.5.0
loguru==0.7.2
tenacity==8.2.3
rich==13.7.0
